requirements.txt
langchain==0.2.16
langchain-core==0.2.38
langchain-openai==0.1.23
langchain-community==0.2.16
langchain-chroma==0.1.4
langchain-text-splitters==0.2.4
langchain-anthropic==0.1.23
langchain-google-genai==1.0.10
pydantic==2.7.4
streamlit==1.37.0
chromadb==0.5.3
pypdf==4.3.1
docx2txt==0.8
tiktoken==0.7.0
pandas==2.2.0
plotly==5.23.0
python-dotenv==1.0.1
sentence-transformers
anthropic>=0.34.0
google-generativeai>=0.7.0
python-docx>=1.1.0